Move filters out of server.js
